modify FAN code
